separate share query component from button
